Move stdlib functions to the separate files.
