+++ library/src/stdlib/abort.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void abort()
+{
+    micros_process_exit(-1);
+}
