+++ library/src/stdlib/atoi.c
@@ -1,0 +1,24 @@
+#include "../stdlib.h"
+
+int32_t atoi(const char *string)
+{
+    size_t len = strlen(string);
+    int32_t value = 0;
+    int32_t position = 1;
+
+    char sign = 1;
+    char lengthOffset = 0;
+    if (*string == '-')
+    {
+        sign = -1;
+        lengthOffset = 1;
+    }
+
+    for (int8_t i = len - 1; i >= lengthOffset; --i)
+    {
+        value += (string[i] - '0') * position;
+        position *= 10;
+    }
+
+    return value * sign;
+}
