+++ library/src/stdlib/exit.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void exit(int status)
+{
+    micros_process_exit(status);
+}
