+++ library/src/stdlib/free.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void free(void *ptr)
+{
+    micros_heap_dealloc(ptr);
+}
