+++ library/src/stdlib/malloc.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void *malloc(size_t size)
+{
+    return micros_heap_alloc(size, 0);
+}
