+++ library/src/stdlib/malloc_align.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void *malloc_align(size_t size, uint32_t align)
+{
+    return micros_heap_alloc(size, align);
+}
