+++ library/src/stdlib/realloc.c
@@ -1,0 +1,6 @@
+#include "../stdlib.h"
+
+void *realloc(void *ptr, size_t size)
+{
+    return micros_heap_realloc(ptr, size);
+}
