+++ library/src/stdlib/srand.c
@@ -1,0 +1,8 @@
+#include "../stdlib.h"
+
+extern unsigned int seed;
+
+void srand(unsigned int new_seed)
+{
+    seed = new_seed;
+}
